requirements.txt
asgiref==3.4.1
bcrypt==3.1.7
blinker==1.4
bootstrap4==0.1.0
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
click==8.0.1
colorama==0.4.3
command-not-found==0.3
confusable-homoglyphs==3.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
defusedxml==0.7.1
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==0.5.0
Django==3.2.7
django-cors-headers==3.10.0
django-filter==21.1
django-heroku==0.3.1
django-registration==3.2
django-registration-redux==2.9
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.20.0
duplicity==0.8.12.0
entrypoints==0.3
Faker==9.3.1
fasteners==0.14.1
Flask==2.0.1
Flask-Hashing==1.1
future==0.18.2
httplib2==0.14.0
idna==2.8
inflection==0.5.1
itsdangerous==2.0.1
itypes==1.2.0
Jinja2==3.0.1
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.3.4
MarkupSafe==2.0.1
monotonic==1.5
netifaces==0.10.4
numpy==1.21.2
oauthlib==3.1.0
olefile==0.46
packaging==21.0
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
psycopg2==2.9.1
pycairo==1.16.2
pycups==1.9.73
PyGObject==3.36.0
PyJWT==2.2.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.5
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-decouple==3.5
python3-openid==3.2.0
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
systemd-python==234
text-unidecode==1.3
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==4.0.0
urllib3==1.25.8
usb-creator==0.3.7
wadllib==1.3.3
Werkzeug==2.0.1
whitenoise==5.3.0
xkit==0.0.0
